move include file handling to separate class
